update with new thoughts, keep old transform and add a default + (#123)
